itracker/requirements.txt
numpy==1.16.4
Pillow==8.2.0
pkg-resources==0.0.0
scipy==1.3.0
six==1.12.0
torch==1.1.0
torchfile==0.1.0
torchvision==0.3.0a0
visdom==0.1.8.9
loguru== 0.5.3
